feat: Created dashboard tabs and pulled top performers data to the dashboard
