+++ src/components/Dashboard/Dashboard.tsx
@@ -1,24 +1,18 @@
 "use client";
 import React, { useState, useEffect, useMemo } from "react";
-import {
-  useAuth,
-  useConference,
-  useUserBracketsByYearId,
-  useProfile,
-  useTeam,
-  useTopPickedTeamsByUser,
-  useUserStatisticsByYearId,
-  useYears,
-} from "@/src/hooks";
+import { useAuth, useProfile, useYears } from "@/src/hooks";
 import {
   LineSpacer,
   LoadingBasketball,
+  MadnessPrepDashboard,
+  Tabs,
+  TabsContent,
+  TabsList,
+  TabsTrigger,
   SelectField,
-  StatBubble,
-  StatList,
-  UserBracketChart,
+  UserStatsDashboard,
 } from "@/src/components";
-import { PRIMARY_COLOR, SECONDARY_COLOR } from "@/src/constants";
+import { SECONDARY_COLOR } from "@/src/constants";
 
 export const Dashboard = () => {
   const [selectedYearId, setSelectedYearId] = useState<string>("");
@@ -27,52 +21,7 @@ export const Dashboard = () => {
   const { data: profileData, isLoading: isLoadingProfile } = useProfile(
     user?.id,
   );
-
   const { data: years, isLoading: isLoadingYears } = useYears();
-  const { data: userBrackets, isLoading: isLoadingUserBracketsByYearId } =
-    useUserBracketsByYearId(user?.id, selectedYearId);
-
-  const userBracketChartData = useMemo(() => {
-    return userBrackets?.map((bracket) => ({
-      score: bracket.score,
-      created_at: bracket.created_at,
-    }));
-  }, [userBrackets]);
-
-  const {
-    isLoadingUserStatisticsByYearId,
-    highScore,
-    lowScore,
-    averageScore,
-    totalBrackets,
-    totalPicks,
-    totalCorrectPicks,
-    correctPickPercentage,
-    totalFirstRoundUpsetsPicked,
-    totalFirstRoundUpsetsCorrect,
-    firstRoundUpsetsCorrectPercentage,
-    lastTenBracketAverage,
-    lastThreeBracketAverage,
-    championCorrectPercentage,
-    finalFourCorrectPercentage,
-    eliteEightCorrectPercentage,
-    sweetSixteenCorrectPercentage,
-    roundTwoCorrectPercentage,
-    roundOneCorrectPercentage,
-    topPickedChampionId,
-    totalTopPickedChampionPicks,
-    topPickedConferenceId,
-    totalTopPickedConferencePicks,
-  } = useUserStatisticsByYearId(user?.id, selectedYearId);
-
-  const { data: topPickedChampion, isLoading: isLoadingTopPickedChampion } =
-    useTeam(topPickedChampionId);
-
-  const { data: topPickedTeams, isLoading: isLoadingTopPickedTeams } =
-    useTopPickedTeamsByUser(user?.id, selectedYearId);
-
-  const { data: topPickedConference, isLoading: isLoadingTopPickedConference } =
-    useConference(topPickedConferenceId);
 
   useEffect(() => {
     if (years?.length && !selectedYearId) {
@@ -85,15 +34,7 @@ export const Dashboard = () => {
     setSelectedYearId(year);
   };
 
-  const isLoading =
-    isLoadingAuth ||
-    isLoadingProfile ||
-    isLoadingYears ||
-    isLoadingUserBracketsByYearId ||
-    isLoadingUserStatisticsByYearId ||
-    isLoadingTopPickedChampion ||
-    isLoadingTopPickedConference ||
-    isLoadingTopPickedTeams;
+  const isLoading = isLoadingAuth || isLoadingProfile || isLoadingYears;
 
   return (
     <div className="w-full py-4 px-8 max-w-screen-2xl mx-auto">
@@ -120,145 +61,23 @@ export const Dashboard = () => {
             selectClassName="w-36 bg-white"
           />
           <LineSpacer lineColor={SECONDARY_COLOR} />
-          <div className="w-full flex sm:flex-row flex-col gap-4">
-            <div className="w-full flex flex-col gap-4">
-              {userBrackets ? (
-                <UserBracketChart
-                  bracketData={userBracketChartData}
-                  lineColor={PRIMARY_COLOR}
-                />
-              ) : (
-                <div className="w-full h-full flex justify-center items-center">
-                  <p>Play some games!</p>
-                </div>
-              )}
-              <div className="w-full flex flex-col gap-4">
-                <div className="w-full grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4 bg-blue-50 rounded-md">
-                  <StatBubble statLabel="High Score" stat={highScore} />
-                  <StatBubble statLabel="Low Score" stat={lowScore} />
-                  <StatBubble statLabel="Average Score" stat={averageScore} />
-                </div>
-                <div className="w-full grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4 bg-blue-50 rounded-md">
-                  <StatBubble
-                    statLabel="Number of Brackets"
-                    stat={totalBrackets}
-                  />
-                  <StatBubble
-                    statLabel="Last Three Average"
-                    stat={lastThreeBracketAverage}
-                    trend={
-                      lastThreeBracketAverage >= averageScore ? "up" : "down"
-                    }
-                  />
-                  <StatBubble
-                    statLabel="Last Ten Average"
-                    stat={lastTenBracketAverage}
-                    trend={
-                      lastTenBracketAverage >= averageScore ? "up" : "down"
-                    }
-                  />
-                </div>
-                <div className="w-full grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4 bg-blue-50 rounded-md">
-                  <StatBubble statLabel="Total Picks" stat={totalPicks} />
-                  <StatBubble
-                    statLabel="Correct Picks"
-                    stat={totalCorrectPicks}
-                  />
-                  <StatBubble
-                    statLabel="Correct Pick %"
-                    stat={correctPickPercentage}
-                    percentage
-                  />
-                </div>
-                <div className="w-full grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4 bg-blue-50 rounded-md">
-                  <StatBubble
-                    statLabel="Picked 1st Round Upsets"
-                    stat={totalFirstRoundUpsetsPicked}
-                  />
-                  <StatBubble
-                    statLabel="Correct 1st Round Upsets"
-                    stat={totalFirstRoundUpsetsCorrect}
-                  />
-                  <StatBubble
-                    statLabel="Correct Upset %"
-                    stat={firstRoundUpsetsCorrectPercentage}
-                    percentage
-                  />
-                </div>
-                <div className="w-full grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4 bg-blue-50 rounded-md">
-                  <StatBubble
-                    statLabel="Champion Pick %"
-                    stat={championCorrectPercentage}
-                    percentage
-                    className="hidden sm:flex"
-                  />
-                  <StatBubble
-                    statLabel="Final Four Pick %"
-                    stat={finalFourCorrectPercentage}
-                    percentage
-                    className="hidden sm:flex"
-                  />
-                  <StatBubble
-                    statLabel="Elite Eight Pick %"
-                    stat={eliteEightCorrectPercentage}
-                    percentage
-                    className="hidden sm:flex"
-                  />
-                  <StatBubble
-                    statLabel="Sweet Sixteen Pick %"
-                    stat={sweetSixteenCorrectPercentage}
-                    percentage
-                    className="hidden sm:flex"
-                  />
-                  <StatBubble
-                    statLabel="Round Two Pick %"
-                    stat={roundTwoCorrectPercentage}
-                    percentage
-                    className="hidden sm:flex"
-                  />
-                  <StatBubble
-                    statLabel="Round One Pick %"
-                    stat={roundOneCorrectPercentage}
-                    percentage
-                    className="hidden sm:flex"
-                  />
-                </div>
-              </div>
-            </div>
-            <div className="flex flex-col gap-4">
-              <StatList
-                title="Top Picked Teams"
-                statLabel="Picks"
-                stats={topPickedTeams?.map((team) => ({
-                  team: team.team_name,
-                  teamLogoRoute: team.team_logo,
-                  stat: team.pick_count,
-                }))}
-              />
-              <StatList
-                title="Top Picked Champion"
-                statLabel="Picks"
-                stats={[
-                  {
-                    team: topPickedChampion?.name,
-                    teamLogoRoute: topPickedChampion?.team_logo,
-                    stat: totalTopPickedChampionPicks ?? 0,
-                  },
-                ]}
-              />
-              <StatList
-                title="Top Picked Conference"
-                statLabel="Picks"
-                stats={[
-                  {
-                    team: topPickedConference?.conference,
-                    teamLogoRoute: topPickedConference?.conference_logo,
-                    stat: totalTopPickedConferencePicks ?? 0,
-                  },
-                ]}
-              />
-            </div>
-          </div>
+          <Tabs
+            className="w-full flex flex-col items-center justify-center"
+            defaultValue="user-stats"
+          >
+            <TabsList>
+              <TabsTrigger value="user-stats">Your Stats</TabsTrigger>
+              <TabsTrigger value="madness-predictions">
+                Madness Prep
+              </TabsTrigger>
+            </TabsList>
+            <TabsContent value="user-stats">
+              <UserStatsDashboard userId={user?.id} yearId={selectedYearId} />
+            </TabsContent>
+            <TabsContent value="madness-predictions">
+              <MadnessPrepDashboard yearId={selectedYearId} />
+            </TabsContent>
+          </Tabs>
         </div>
       )}
     </div>
