+++ src/components/Dashboard/index.ts
@@ -1,6 +1,8 @@
 export * from "./Dashboard";
+export * from "./MadnessPrepDashboard";
 export * from "./StatBubble";
 export * from "./StatList";
 export * from "./StatListItem";
 export * from "./TopStatListItem";
 export * from "./UserBracketChart";
+export * from "./UserStatsDashboard";
